upd
-boton descargar en comparador de credito

+++ Recursos/Scripts/AgregarEliminarFilas.js
@@ -15,11 +15,11 @@ var cont=0;
 var montoTotal=0;
 var id_fila_selected=[];
 function agregar(){
-    cont++;
-    montoTotal = parseInt(montoTotal) + parseInt($("#monto").val());
+	cont++;
+	montoTotal = parseFloat(montoTotal) + parseFloat($("#monto").val());
 	var fila='<tr class="selected" id="fila'+cont+'" onclick="seleccionar(this.id);"><td>'+cont+
-			'</td><td>'+ $("#inst").val() +
-			'</td><td>'+ $("#monto").val() +'</td></tr>';
+	'</td><td>'+ $("#inst").val() +
+	'</td><td>'+ $("#monto").val() +'</td></tr>';
 	$('#tabla').append(fila);
 	reordenar();
 
@@ -29,8 +29,8 @@ function agregar(){
 
 function calcular (){
 	var interes1 = Math.floor((Math.random() * (17 - 14)) + 14)/100;
-   	var interes2 = Math.floor((Math.random() * (17 - 14)) + 14)/100;
-    var arreycuotas = [24,36,48];
+	var interes2 = Math.floor((Math.random() * (17 - 14)) + 14)/100;
+	var arreycuotas = [24,36,48];
 	var periodo = 12;
 	var numeroCuotas1 = arreycuotas[Math.floor((Math.random() * (3 - 0)) + 0)];
 	var numeroCuotas2 = arreycuotas[Math.floor((Math.random() * (3 - 0)) + 0)];
@@ -49,108 +49,109 @@ function calcular (){
 
 	if($("#sd").val()>cuota1 && $("#sd").val()>cuota2){
 
-	    $("#montoA").val(montoTotal);
+		$("#montoA").val(montoTotal);
+		
 
-	    renderizarTablaFrancesa1({
-	        cuota: cuota1,
-	        valorPrestamo: montoTotal,
-	        numeroCuotas: numeroCuotas1,
-	        interes: interesCuota1,
-	        intbx: interes1,
-	        periodo: periodo
-	    });
+		renderizarTablaFrancesa1({
+			cuota: cuota1,
+			valorPrestamo: montoTotal,
+			numeroCuotas: numeroCuotas1,
+			interes: interesCuota1,
+			intbx: interes1,
+			periodo: periodo
+		});
 
-	    renderizarTablaFrancesa2({
-	        cuota: cuota2,
-	        valorPrestamo: montoTotal,
-	        numeroCuotas: numeroCuotas2,
-	        interes: interesCuota2,
-	        intby: interes2,
-	        periodo: periodo
-	    });
+		renderizarTablaFrancesa2({
+			cuota: cuota2,
+			valorPrestamo: montoTotal,
+			numeroCuotas: numeroCuotas2,
+			interes: interesCuota2,
+			intby: interes2,
+			periodo: periodo
+		});
 
-	    var html = '<button onclick="HTMLtoPDF2()" type="button" class="btn btn-success">Descargar PDF</button>';
+		var html = '<button onclick="HTMLtoPDF2()" type="button" class="btn btn-success">Descargar PDF</button>';
 		$('#downloadbutton').html(html);
 	}else {
 		alert("El monto del salario débe ser mayor a la cuota a pagar, ya que deducción se hace por planilla.");
-		};
+	};
 }
 
 function renderizarTablaFrancesa1(object) {
-    $('#tablaC1').html("");
+	$('#tablaC1').html("");
 
-   	var html = '<h4>Banco X, Ofrece:</h4><p>Interés: '+ (object.intbx*100).toFixed(2) +'%</p>';
+	var html = '<h4>Banco X, Ofrece:</h4><p>Interés: '+ (object.intbx*100).toFixed(2) +'%</p>';
 	html += '<p>Número de Cuotas: '+ object.numeroCuotas +'</p>';
 	html += '<table class="table"><thead class="thead-dark"><tr><th scope="col">Períodos</th><th scope="col">Cuotas</th><th scope="col">Interéses</th><th scope="col">Amortización</th><th scope="col">Capital Pendiente</th></tr></thead><tbody>';
 	html += '<tr><th scope="row">0</th><td></td><td></td><td></td><td>' + object.valorPrestamo + '</td></tr>';
 
 	var interesTemp = 0;
 	var interesAcumulado = 0;
-	    
+
 	var labels = [0];
 	var data = {
-	    capital: [object.valorPrestamo],
-	    interes: [0],
-	    amortizacion: [0],
-	    cuota: [0]
+		capital: [object.valorPrestamo],
+		interes: [0],
+		amortizacion: [0],
+		cuota: [0]
 	};
 
 	for (let i = 0; i < object.numeroCuotas; i++) {
-	    interesTemp = object.interes * object.valorPrestamo / object.periodo;
-	    object.valorPrestamo -= object.cuota - interesTemp;
-	    interesAcumulado += interesTemp;
-	        
-	    labels.push(i + 1);
-	    data.capital.push(object.valorPrestamo);
-	    data.interes.push(interesTemp);
-	    data.amortizacion.push(object.cuota1 - interesTemp);
-	    data.cuota.push(object.cuota);
-	        
-	    html += '<tr><th scope="row">' + (i + 1) + '</th><td>' + object.cuota.toFixed(2) + '</td><td>' + interesTemp.toFixed(2) + '</td><td>' + (object.cuota - interesTemp).toFixed(2) + '</td><td>' + object.valorPrestamo.toFixed(2) + '</td></tr>';
+		interesTemp = object.interes * object.valorPrestamo / object.periodo;
+		object.valorPrestamo -= object.cuota - interesTemp;
+		interesAcumulado += interesTemp;
+
+		labels.push(i + 1);
+		data.capital.push(object.valorPrestamo);
+		data.interes.push(interesTemp);
+		data.amortizacion.push(object.cuota1 - interesTemp);
+		data.cuota.push(object.cuota);
+
+		html += '<tr><th scope="row">' + (i + 1) + '</th><td>' + object.cuota.toFixed(2) + '</td><td>' + interesTemp.toFixed(2) + '</td><td>' + (object.cuota - interesTemp).toFixed(2) + '</td><td>' + object.valorPrestamo.toFixed(2) + '</td></tr>';
 	}
 
 	html += '<tr><th scope="row">Totales:</th><td>' + (object.cuota * object.numeroCuotas).toFixed(2) + '</td><td>' + interesAcumulado.toFixed(2) + '</td><td>' + (object.cuota * object.numeroCuotas - interesAcumulado).toFixed(2) + '</td><td></td></tr></tbody></table>';
 
-    $('#tablaC1').html(html);
+	$('#tablaC1').html(html);
 
 }
 
 function renderizarTablaFrancesa2(object) {
-    $('#tablaC2').html("");
+	$('#tablaC2').html("");
 
-   	var html = '<h4>Banco Y, Ofrece:</h4><p>Interés: '+ (object.intby*100).toFixed(2) +'%</p>';
+	var html = '<h4>Banco Y, Ofrece:</h4><p>Interés: '+ (object.intby*100).toFixed(2) +'%</p>';
 	html += '<p>Número de Cuotas: '+ object.numeroCuotas +'</p>';
 	html += '<table class="table"><thead class="thead-dark"><tr><th scope="col">Períodos</th><th scope="col">Cuotas</th><th scope="col">Interéses</th><th scope="col">Amortización</th><th scope="col">Capital Pendiente</th></tr></thead><tbody>';
 	html += '<tr><th scope="row">0</th><td></td><td></td><td></td><td>' + object.valorPrestamo + '</td></tr>';
 
 	var interesTemp = 0;
 	var interesAcumulado = 0;
-	    
+
 	var labels = [0];
 	var data = {
-	    capital: [object.valorPrestamo],
-	    interes: [0],
-	    amortizacion: [0],
-	    cuota: [0]
+		capital: [object.valorPrestamo],
+		interes: [0],
+		amortizacion: [0],
+		cuota: [0]
 	};
 
 	for (let i = 0; i < object.numeroCuotas; i++) {
-	    interesTemp = object.interes * object.valorPrestamo / object.periodo;
-	    object.valorPrestamo -= object.cuota - interesTemp;
-	    interesAcumulado += interesTemp;
-	        
-	    labels.push(i + 1);
-	    data.capital.push(object.valorPrestamo);
-	    data.interes.push(interesTemp);
-	    data.amortizacion.push(object.cuota1 - interesTemp);
-	    data.cuota.push(object.cuota);
-	        
-	    html += '<tr><th scope="row">' + (i + 1) + '</th><td>' + object.cuota.toFixed(2) + '</td><td>' + interesTemp.toFixed(2) + '</td><td>' + (object.cuota - interesTemp).toFixed(2) + '</td><td>' + object.valorPrestamo.toFixed(2) + '</td></tr>';
+		interesTemp = object.interes * object.valorPrestamo / object.periodo;
+		object.valorPrestamo -= object.cuota - interesTemp;
+		interesAcumulado += interesTemp;
+
+		labels.push(i + 1);
+		data.capital.push(object.valorPrestamo);
+		data.interes.push(interesTemp);
+		data.amortizacion.push(object.cuota1 - interesTemp);
+		data.cuota.push(object.cuota);
+
+		html += '<tr><th scope="row">' + (i + 1) + '</th><td>' + object.cuota.toFixed(2) + '</td><td>' + interesTemp.toFixed(2) + '</td><td>' + (object.cuota - interesTemp).toFixed(2) + '</td><td>' + object.valorPrestamo.toFixed(2) + '</td></tr>';
 	}
 
 	html += '<tr><th scope="row">Totales:</th><td>' + (object.cuota * object.numeroCuotas).toFixed(2) + '</td><td>' + interesAcumulado.toFixed(2) + '</td><td>' + (object.cuota * object.numeroCuotas - interesAcumulado).toFixed(2) + '</td><td></td></tr></tbody></table>';
 
-    $('#tablaC2').html(html);
+	$('#tablaC2').html(html);
 
 }
